[zen-27] Renamed interfaces in Zend\Acl Removed some @uses
