+++ library/Zend/Acl/Assertion/AssertionInterface.php
@@ -18,18 +18,16 @@
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
 
-namespace Zend\Acl;
+namespace Zend\Acl\Assertion;
 
+use Zend\Acl;
 /**
- * @uses       Zend\Acl\Acl
- * @uses       Zend\Acl\Resource
- * @uses       Zend\Acl\Role
  * @category   Zend
  * @package    Zend_Acl
  * @copyright  Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
-interface Assertion
+interface AssertionInterface
 {
     /**
      * Returns true if and only if the assertion conditions are met
@@ -38,11 +36,11 @@ interface Assertion
      * $role, $resource, or $privilege parameters are null, it means that the query applies to all Roles, Resources, or
      * privileges, respectively.
      *
-     * @param  Acl      $acl
-     * @param  Role     $role
-     * @param  Resource $resource
-     * @param  string   $privilege
+     * @param  Acl                           $acl
+     * @param  Role\RoleInterface            $role
+     * @param  Resource\ResourceInterface    $resource
+     * @param  string                        $privilege
      * @return boolean
      */
-    public function assert(Acl $acl, Role\RoleInterface $role = null, Resource\ResourceInterface $resource = null, $privilege = null);
+    public function assert(Acl\Acl $acl, Acl\Role\RoleInterface $role = null, Acl\Resource\ResourceInterface $resource = null, $privilege = null);
 }
