+++ library/Zend/Acl/Exception/ExceptionInterface.php
@@ -21,7 +21,6 @@
 namespace Zend\Acl\Exception;
 
 /**
- * @uses       Zend\Exception
  * @category   Zend
  * @package    Zend_Acl
  * @copyright  Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
