+++ library/Zend/Acl/Resource/GenericResource.php
@@ -23,7 +23,6 @@ namespace Zend\Acl\Resource;
 use Zend\Acl\Resource\ResourceInterface;
 
 /**
- * @uses       \Zend\Acl\Resource\ResourceInterface
  * @category   Zend
  * @package    Zend_Acl
  * @copyright  Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
