+++ library/Zend/Acl/Role/GenericRole.php
@@ -23,7 +23,6 @@ namespace Zend\Acl\Role;
 use Zend\Acl\Role\RoleInterface;
 
 /**
- * @uses       Zend\Acl\Role
  * @category   Zend
  * @package    Zend_Acl
  * @copyright  Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
