+++ library/Zend/Acl/Role/Registry.php
@@ -26,8 +26,6 @@ use Zend\Acl,
 
 
 /**
- * @uses       Zend\Acl\Role
- * @uses       Zend\Acl\Role\Exception
  * @category   Zend
  * @package    Zend_Acl
  * @copyright  Copyright (c) 2005-2012 Zend Technologies USA Inc. (http://www.zend.com)
