+++ tests/Zend/Acl/TestAsset/AssertionZF7973.php
@@ -2,10 +2,10 @@
 
 namespace ZendTest\Acl\TestAsset;
 
-use Zend\Acl\Assertion,
+use Zend\Acl\Assertion\AssertionInterface,
     Zend\Acl;
 
-class AssertionZF7973 implements Assertion {
+class AssertionZF7973 implements AssertionInterface {
     public function assert(Acl\Acl $acl, Acl\Role\RoleInterface $role = null, Acl\Resource\ResourceInterface $resource = null, $privilege = null)
     {
         if($privilege != 'privilege') {
