+++ tests/Zend/Acl/TestAsset/UseCase1/UserIsBlogPostOwnerAssertion.php
@@ -2,10 +2,10 @@
 
 namespace ZendTest\Acl\TestAsset\UseCase1;
 
-use Zend\Acl\Assertion,
+use Zend\Acl\Assertion\AssertionInterface,
     Zend\Acl as ZendAcl;
 
-class UserIsBlogPostOwnerAssertion implements Assertion
+class UserIsBlogPostOwnerAssertion implements AssertionInterface
 {
 
     public $lastAssertRole = null;
